requirements.txt
aiohttp
aiosignal
annotated-types
anthropic
anyio
asttokens
async-timeout
attrs
black
certifi
charset-normalizer
click
comm
databind.core
databind.json
debugpy
decorator
Deprecated
diskcache
distro
docspec
docspec-python
docstring-parser
exceptiongroup
executing
filelock
FLAML
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
idna
importlib-metadata
ipykernel
ipython
jedi
Jinja2
joblib
jupyter_client
jupyter_core
MarkupSafe
matplotlib-inline
mpmath
multidict
mypy-extensions
nanoid
nest-asyncio
networkx
nr-date
nr-stream
nr.util
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
openai
packaging
pandas
parso
pathspec
pexpect
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pydantic
pydantic_core
pydoc-markdown
Pygments
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
requests
scikit-learn
scipy
six
sniffio
stack-data
sympy
termcolor
threadpoolctl
tokenizers
tomli
tomli_w
torch
tornado
tqdm
traitlets
triton
typeapi
typing_extensions
tzdata
urllib3
watchdog
wcwidth
wrapt
yapf
yarl
zipp
